Initial commit
First working version of a REST extractor, with a very limitied set of
capabilities:

 * No auth handling
 * Only supports writing Events
 * No throttling
 * No incremental load
 * No timed runs
 * No support for paginated results

All of these are planned features for an initial release, but is kept
out of the initial commit to keep the scope of this commit down.

The idea for the REST extension is to allow users to write a decorator
for how the HTTP call should be made, and a function to handle adapting
the resulting data to the CDF data model, like so:

``` python
extractor = RestExtractor(...)

@extractor.get("https://api.source-system.com/events", response_type=EventsList)
def get_events(events: EventsList) -> Iterable[Event]:
    for event in events.items:
        yield Event(
            external_id=f"testy-{event.id}",
            start_time=event.start_time,
            end_time=event.end_time
        )

with extractor:
    extractor.run()
```

Anyone familiar with Flask, FastAPI or other smilar frameworks will see
where the inspiration comes from.