Add RAW capabilites (#4)
Add a RAW upload queue and allow handle functions to return rows.

Since we need to know the db/table (and those are not attributes in the
Row class from the SDK), I think we need to define our own types. To
avoid possible confusion, and in case we need to do similar things for
other resource types, I added a types module, which contains type
aliases for those types where we just use the SDK types directly. That
way, users can always import from our type module and be safe.