Change the way queries are executed (#32)
Also add a few tests.

This fundamentally changes how the extractor executes queries. Instead of running each registered query in a separate thread, run a common executor that waits on a priority queue with the scheduled runs.

 - Retrieve HTTP call from priority queue, the priority is equal to the time it is expected to execute, or 0 if it should be executed as soon as possible.
 - If the call has a time set to greater than 0, we begin waiting on the queue.
 - Since new calls added to the queue may have lower priority than the one we are currently waiting for, we have to check if we should wait before moving on.
 - If there is no call in the queue, we may abort only when there also is no running call. Rather than doing some crazy gating here, I just ended up with an integer behind a lock. It should be sound. When we are not waiting for a call we set a default timeout (1 second), to allow checking for whether all calls have terminated without anything being added to the queue.

This now technically allows adding new calls to the execution queue while the extractor is running, though doing so from outside of another call would be very prone to unexpected behavior.